+++ src/scripts/main.js
@@ -1,285 +1,10 @@
-var $ = require('jquery');
+//const $ = require('jquery');
+const hamburger = require('./modules/hamburger');
+const welcome = require('./modules/welcome');
+const googleMap = require('./modules/map');
+const skills = require('./modules/skills');
 
-$(document).ready(function () {
-    
-    $(function(){
-        $('.intro__btn-auth').on('click',function (e) {
-            e.preventDefault();
-            var elem = $(e.target),
-                mainIntro = elem.closest('.main--intro');
-
-            mainIntro.addClass('active');
-        });
-
-        $('.form__button--home').on('click',function (e) {
-            e.preventDefault();
-            var elem = $(e.target),
-                mainIntro = elem.closest('.main--intro');
-
-            mainIntro.removeClass('active');
-        });
-    });
-
-    $(function(){
-        var hamburger = $('.hamburger');
-        $('.hamburger__btn').on('click',function (e) {
-            e.preventDefault();
-            hamburger.toggleClass('active');
-        });
-    });
-
-    $(function(){
-        $('.circle').on('mouseover',function (e) {
-            e.preventDefault();
-            var elem = $(e.target),
-            circleColored = elem.find('.circle__second'),
-            percent = circleColored.attr('data-circle');
-            strokeDasharray = 2 * Math.PI * 45 * percent / 100;
-            circleColored.css('stroke-dasharray', strokeDasharray + ' 282.6');
-        });
-
-    });
-
-});
-
-window.initMap = function () {
-    var centerPoint = {lat: 55.783745, lng: 37.595038};
-    var markerPoint = {lat: 55.800261, lng: 37.532870};
-    var map = new google.maps.Map(document.getElementById('map'), {
-    zoom: 13,
-    center: centerPoint,
-    styles: [
-        {
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#f5f5f5"
-            }
-          ]
-        },
-        {
-          "elementType": "labels.icon",
-          "stylers": [
-            {
-              "visibility": "off"
-            }
-          ]
-        },
-        {
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#616161"
-            }
-          ]
-        },
-        {
-          "elementType": "labels.text.stroke",
-          "stylers": [
-            {
-              "color": "#f5f5f5"
-            }
-          ]
-        },
-        {
-          "featureType": "administrative.land_parcel",
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#bdbdbd"
-            }
-          ]
-        },
-        {
-          "featureType": "landscape",
-          "elementType": "geometry.fill",
-          "stylers": [
-            {
-              "color": "#ffffff"
-            }
-          ]
-        },
-        {
-          "featureType": "poi",
-          "stylers": [
-            {
-              "visibility": "off"
-            }
-          ]
-        },
-        {
-          "featureType": "poi",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#eeeeee"
-            }
-          ]
-        },
-        {
-          "featureType": "poi",
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#757575"
-            }
-          ]
-        },
-        {
-          "featureType": "poi.park",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#e5e5e5"
-            }
-          ]
-        },
-        {
-          "featureType": "poi.park",
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#9e9e9e"
-            }
-          ]
-        },
-        {
-          "featureType": "road",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#ffffff"
-            }
-          ]
-        },
-        {
-          "featureType": "road",
-          "elementType": "geometry.fill",
-          "stylers": [
-            {
-              "color": "#e3e3e3"
-            }
-          ]
-        },
-        {
-          "featureType": "road",
-          "elementType": "geometry.stroke",
-          "stylers": [
-            {
-              "color": "#e3e3e3"
-            }
-          ]
-        },
-        {
-          "featureType": "road",
-          "elementType": "labels.icon",
-          "stylers": [
-            {
-              "visibility": "off"
-            }
-          ]
-        },
-        {
-          "featureType": "road",
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#adadad"
-            }
-          ]
-        },
-        {
-          "featureType": "road.highway",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#dadada"
-            }
-          ]
-        },
-        {
-          "featureType": "road.highway",
-          "elementType": "geometry.fill",
-          "stylers": [
-            {
-              "color": "#d3d3d3"
-            }
-          ]
-        },
-        {
-          "featureType": "transit",
-          "stylers": [
-            {
-              "visibility": "off"
-            }
-          ]
-        },
-        {
-          "featureType": "transit.line",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#e5e5e5"
-            }
-          ]
-        },
-        {
-          "featureType": "transit.station",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#eeeeee"
-            }
-          ]
-        },
-        {
-          "featureType": "water",
-          "elementType": "geometry",
-          "stylers": [
-            {
-              "color": "#c9c9c9"
-            }
-          ]
-        },
-        {
-          "featureType": "water",
-          "elementType": "geometry.fill",
-          "stylers": [
-            {
-              "color": "#61dac9"
-            }
-          ]
-        },
-        {
-          "featureType": "water",
-          "elementType": "labels",
-          "stylers": [
-            {
-              "visibility": "off"
-            }
-          ]
-        },
-        {
-          "featureType": "water",
-          "elementType": "labels.text.fill",
-          "stylers": [
-            {
-              "color": "#9e9e9e"
-            }
-          ]
-        }
-      ]
-    });
-    var marker = new google.maps.Marker({
-        position: markerPoint,
-        map: map,
-        title: "Здесь может быть удобно встретиться",
-        icon: {
-            url:'../img/map_marker.png',
-            size: new google.maps.Size(30, 40),
-            origin: new google.maps.Point(0, 0),
-            anchor: new google.maps.Point(0, 40),
-            scale: 2
-        }
-    });
-};
+hamburger();
+welcome();
+googleMap();
+skills();
