+++ src/scripts/modules/hamburger.js
@@ -1,0 +1,9 @@
+var $ = require('jquery');
+
+module.exports = $(function(){
+    var hamburger = $('.hamburger');
+    $('.hamburger__btn').on('click',function (e) {
+        e.preventDefault();
+        hamburger.toggleClass('active');
+    });
+});
