+++ src/styles/components/auth.scss
@@ -1,24 +1,18 @@
 .auth {
     position: absolute;
-    //top: 50%;
-    //left: 50%;
-    //transform: translate(-50%,-50%);
     width: 100%;
     height: 100%;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    //position: relative;
-    //display: none;
-
     backface-visibility: hidden;
     transform: rotateY(180deg);   
 
 }
 
 .auth-form {
-    width: 380px;
+    width: 100%;
     border-radius: 5px;
     display: flex;
     flex-direction: column;
