+++ src/styles/components/contact-form.scss
@@ -19,7 +19,13 @@
         top: 50%;
         left: 50%;
         transform: translate(-50%, -50%);
-        background: url('../img/bg/bg_comments.jpg') 50% 67% no-repeat;
+        background: url('../img/bg/bg_comments.jpg') 50% 50% no-repeat;
         filter: blur(25px);
+    }
+
+    @include tablets {
+        width: 348px;
+        margin-bottom: 100px;
+
     }
 }
