+++ src/styles/components/footer.scss
@@ -12,6 +12,10 @@
     &:last-child {
         border-color: transparent;
     }
+
+    @include tablets {
+        height: 117px;
+    }
 }
 
 .footer__row .container{
@@ -20,7 +24,13 @@
     align-items: center;
     padding: 15px 20px;
 
-
+    @include tablets {
+        height: 100%;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        padding: 0px 20px;
+    }
 }
 
 .footer__info-text, .footer__copyright {
@@ -32,8 +42,19 @@
     @include desktop-large {
         width: 30%;
     }
+
+    @include tablets {
+        width: 100%;
+        text-align: center;
+        
+    }
 }
   
 .footer__copyright {
     text-align: right;
+
+    @include tablets {
+        text-align: center;
+        margin-top: 15px;
+    }
 }
