+++ src/styles/components/hamburger.scss
@@ -11,6 +11,11 @@
     right: 30px;
     z-index: 200;
 
+    @include tablets {
+        top: 27px;
+        right: 24px;
+    }
+
 }
 
 .hamburger__icon {
@@ -64,11 +69,32 @@
     &:last-child {
         margin-bottom: 0;
     }
+
+    @include tablets {
+        margin-bottom: 66px;
+    }
 }
 
 .hamburger__link {
     color: $white;
     text-decoration: none;
     @include title-decor($white,40px);
+    transition: all 0.5s;
+
+    @include tablets {
+        @include title-decor($white,30px);
+    }
+
+
+}
+
+.hamburger__item:hover {
+    & .hamburger__link:before{
+        opacity: 1;
+    }
+
+    & .hamburger__link:after{
+        opacity: 1;
+    }
 
 }
