+++ src/styles/components/intro.scss
@@ -19,6 +19,7 @@
     justify-content: center;
     border: 0px solid transparent;
     height: 100%;
+    width: 100%;
 
 
 }
@@ -53,6 +54,12 @@
     transform: translate(-50%, 0%);
     color: $white;
     opacity: 0.7;
+    text-align: center;
+    width: 100%;
+
+    @include tablets {
+        font-size: 14px;
+    }
 }
