+++ src/styles/components/map.scss
@@ -16,3 +16,17 @@
     height: 100%;
     width: 100%;
 }
+
+a[href^="http://maps.google.com/maps"]{
+  display:none !important
+}
+a[href^="https://maps.google.com/maps"]{
+  display:none !important
+}
+
+.gmnoprint a, .gmnoprint span, .gm-style-cc {
+    display:none;
+}
+.gmnoprint div {
+    background:none !important;
+}
