+++ src/styles/components/skills.scss
@@ -22,11 +22,20 @@
     flex-direction: column;
     align-items: flex-start;
 
+    @include tablets {
+        width: 60%;
+        align-items: center;
+        
+    }
 
 }
 
 .skills__text {
     margin-bottom: 25px;
+
+    @include tablets {
+        margin-bottom: 18px;
+    }
 }
 
 .skills__group {
@@ -34,6 +43,17 @@
     
     &:last-child {
         margin: 0;
+
+        @include tablets {
+            margin-bottom: 50px;
+        }
+    }
+
+    @include tablets {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin-bottom: 42px;
     }
 }
 
@@ -41,4 +61,8 @@
     @include roboto(500);
     font-size: 28px;
     margin-bottom: 10px;
+
+    @include tablets {
+        font-size: 21px;
+    }
 }
