+++ src/styles/components/subtitle.scss
@@ -12,9 +12,16 @@
         text-align: center;
     }
 
+    @include tablets {
+        font-size: 30px;
+    }
+
     &--about {
         margin-bottom: 30px;
 
+        @include tablets {
+            margin-bottom: 5px;
+        }
 
     }
 
@@ -23,6 +30,13 @@
         margin-bottom: 30px;
         width: 320px;
         align-self: center;
+        text-align: center;
+
+        @include tablets {
+            margin-top: 37px;
+            margin-bottom: 3px;
+            
+        }
 
     }
 
@@ -38,11 +52,19 @@
         &:after {
             display: none;
         }
+
+        @include tablets {
+            margin-bottom: 5px;
+        }
     }
 
     &--works {
         margin-bottom: 30px;
         text-align: center;
+
+        @include tablets {
+            margin-bottom: 13px;
+        }
     }
 
     &--contactform {
@@ -50,6 +72,10 @@
         margin-bottom: 30px;
         margin-top: 30px;
         display: flex;
+
+        @include tablets {
+            margin-bottom: 11px;
+        }
     }
 
 }
