+++ src/styles/components/title.scss
@@ -31,11 +31,21 @@
     margin: 0;
     text-align: center;
 
+    @include tablets {
+        font-size: 49px;
+        @include title-decor($black,30px);
+    }
+
     &--about {
         margin-bottom: 40px;
 
         @include desktop-large {
             margin-right: 60px;
+        }
+
+        @include tablets {
+            margin-bottom: 30px;
+            margin-top: 30px;
         }
     }
 
@@ -49,9 +59,16 @@
 
     &--comments {
         width: 380px;
-        margin-top: 65px;
-        margin-bottom: 65px;
+        margin-top: 60px;
+        margin-bottom: 60px;
         line-height: 66px;
+
+        @include tablets {
+            width: 472px;
+            margin-top: 32px;
+            margin-bottom: 32px;
+
+        }
     }
 }
 
@@ -66,6 +83,10 @@
 
     &--about {
         fill: #6e7171;
+
+        @include tablets {
+            top: 0%;
+        }
     }
 
     &--blog {
@@ -77,6 +98,10 @@
         transform: translate(-50%, -20%);
         width: 50%;
         fill: #6e7171;
+
+        @include tablets {
+            top: 10%;
+        }
     }
 
     &--comments {
