+++ src/templates/pages/index.pug
@@ -22,21 +22,24 @@ block content
                             var menuItems = [
                                 {
                                 title: "Мои работы", 
-                                link: "./works.html"
+                                link: "./works.html",
+                                mod: "works"
                                 },
                                 {
                                 title: "Обо мне", 
-                                link: "./about.html"
+                                link: "./about.html",
+                                mod: "about"
                                 },
                                 {
                                 title: "Блог", 
-                                link: "./blog.html"
+                                link: "./blog.html",
+                                mod: "blog"
                                 }
                             ]
 
-                        ul.menu
+                        ul.menu__list
                             each item in menuItems
-                                li.menu__item
+                                li.menu__item(class=`menu__item--${item.mod}`)
                                     a(href=`${item.link}`).menu__link #{item.title}
             section.auth
                 .auth-form
