+++ src/templates/pages/works.pug
@@ -86,7 +86,6 @@ block content
                 ul.comments__list
                     each item in commentsList
                         li.comments__item
-                            img(src="img/quote.png").comments__quote-icon
                             .comments__text
                                 p #{item.comment}
                             .comments__photo
